refactor: users hooks moved to sep. folder ♻️
